Container name shouldn't be hardcoded as it depends on the NAME parameter
Currently, the container name is hardcoded as "rails-pgsql-persistent" which
is the default value. However, if user changes the default value in NAME, then
the container name is not matched which causes the failure due to the container
can't be found.

The container name should n't be hardcoded as it should be based on the NAME
parameter that is provided and modified by user.

Signed-off-by: Vu Dinh <vdinh@redhat.com>