Fix nokogiri dependencies
https://github.com/sclorg/rails-ex/pull/132 removed the mini_portile2
dependency from nokogiri.  On x86-linux this still seemed to work, but
on other architectures for which nokogiri has to be compiled from source,
this caused the build to fail:

Downloading nokogiri-1.11.2 revealed dependencies not in the API or the lockfile
(mini_portile2 (~> 2.5.0)).
Either installing with `--full-index` or running `bundle update nokogiri` should
fix the problem.